fix: sort component responsivity
Refactor: remove duplicated sort components.
