+++ apps/ui/components/form/index.ts
@@ -1,15 +1,6 @@
-import Checkbox from "./Checkbox";
-import DateRangePicker from "./DateRangePicker";
-import Dropdown from "./Dropdown";
-import ScrollIntoViewWithFocus from "./ScrollIntoViewWithFocus";
-import SearchLabel from "@/components/form/SearchLabel";
-import Sorting from "./Sorting";
-
-export {
-  Checkbox,
-  DateRangePicker,
-  Dropdown,
-  ScrollIntoViewWithFocus,
-  SearchLabel,
-  Sorting,
-};
+export { default as Checkbox } from "./Checkbox";
+export { default as DateRangePicker } from "./DateRangePicker";
+export { default as Dropdown } from "./Dropdown";
+export { default as ScrollIntoViewWithFocus } from "./ScrollIntoViewWithFocus";
+export { default as SearchLabel } from "@/components/form/SearchLabel";
+export { Sorting } from "./Sorting";
